move EventList component one directory up
